+++ pyproject.toml
@@ -2,6 +2,7 @@
 requires = [
   "setuptools>=65.0",
   "setuptools-scm[toml]>=8.0",
+  "mypy>=1.15",
 ]
 build-backend = "setuptools.build_meta"
