+++ src/pluggy/_manager.py
@@ -5,7 +5,7 @@ from collections.abc import Mapping
 from collections.abc import Sequence
 import inspect
 import types
-from typing import Any
+from typing import Any, TypeAlias
 from typing import Callable
 from typing import cast
 from typing import Final
@@ -13,6 +13,7 @@ from typing import TYPE_CHECKING
 import warnings
 
 from . import _tracing
+from ._hookrelay import HookRelay
 from ._callers import _multicall
 from ._hooks import _HookImplFunction
 from ._hooks import _Namespace
@@ -21,7 +22,6 @@ from ._hooks import _SubsetHookCaller
 from ._hooks import HookCaller
 from ._hooks import HookImpl
 from ._hooks import HookimplOpts
-from ._hooks import HookRelay
 from ._hooks import HookspecOpts
 from ._hooks import normalize_hookimpl_opts
 from ._result import Result
@@ -32,8 +32,8 @@ if TYPE_CHECKING:
     import importlib.metadata
 
 
-_BeforeTrace = Callable[[str, Sequence[HookImpl], Mapping[str, Any]], None]
-_AfterTrace = Callable[[Result[Any], str, Sequence[HookImpl], Mapping[str, Any]], None]
+_BeforeTrace: TypeAlias = Callable[[str, Sequence[HookImpl], Mapping[str, Any]], None]
+_AfterTrace: TypeAlias = Callable[[Result[Any], str, Sequence[HookImpl], Mapping[str, Any]], None]
 
 
 def _warn_for_function(warning: Warning, function: Callable[..., object]) -> None:
